% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothOR.R
\name{smoothOR}
\alias{smoothOR}
\title{Smooth OR}
\usage{
smoothOR(data, response = NULL, formula = NULL, gamfit)
}
\arguments{
\item{data}{A data frame containing the dataset for analysis.}

\item{response}{The response variable in the data frame. Required when fitting a new model (modelfit = "FALSE")}

\item{formula}{The formula describing the model. Required when fitting a new model (modelfit = "FALSE").}

\item{gamfit}{A pre-fitted GAM object. When provided, modelfit is set to "TRUE" and the function processes this object.}
}
\value{
An object of class "OR" containing the processed data and analysis results.

The returned object includes relevant information from the fitted Generalized Additive Model (GAM) or
analysis of the provided GAM object. It provides insights into odds ratios and their analysis.
Users can further explore and interpret the results using appropriate methods for class "OR".
}
\description{
fits a Generalized Additive Model (GAM) and creating a resulting object
containing the model outcomes, related to odds ratio analysis.
Here's a step-by-step description of what the function does:
}
\seealso{
\code{\link{gam}}, \code{\link{summary}}, \code{\link{plot}}
}
\keyword{internal}
