% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICc.R
\name{AICc}
\alias{AICc}
\title{Akaike's Information Criterion corrected}
\usage{
AICc(object)
}
\arguments{
\item{object}{This object has to be of class 'GAM' or 'gam'.}
}
\value{
Returns the AICc coefficient.
}
\description{
This function calculates the corrected Akaike Information Criterion (AICc)
for a given statistical model. AICc is a modification of the AIC that
provides a correction for small sample sizes, which can lead to overfitting
in model selection. AICc = AIC + 2*p(p+1)/(n-p-1)
}
\examples{
library(mgcv)
set.seed(123)
n <- 100
x <- rnorm(n)
y <- sin(x) + rnorm(n, sd = 0.2)
gam_fit <- gam(y ~ s(x))
AICc(gam_fit)

}
\keyword{internal}
