% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfgam.R
\name{dfgam}
\alias{dfgam}
\title{Selection of the optimal degrees of freedom}
\usage{
dfgam(
  response,
  nl.predictors,
  other.predictors = NULL,
  smoother = "s",
  method = "AIC",
  data,
  step = NULL
)
}
\arguments{
\item{response}{The name of the response variable in the dataset.}

\item{nl.predictors}{A character vector containing the names of the predictors for which optimal degrees of freedom need to be determined.}

\item{other.predictors}{A character vector containing the names of other predictors to be included in the model (optional).}

\item{smoother}{The type of smoother to be used. Currently, only "s" (thin plate regression spline) is supported.}

\item{method}{The information criterion to be used for selection. Possible values: "AIC", "AICc", or "BIC".}

\item{data}{The dataset containing the variables.}

\item{step}{The number of steps to search for optimal degrees of freedom (optional, default is 3).}
}
\value{
A list containing the fitted model, a matrix of selected degrees of freedom, the chosen information criterion method,
the names of predictors for which degrees of freedom were determined, and other predictors included in the model (if any).
}
\description{
This function performs the selection of optimal degrees of freedom for smoothing terms
in a Generalized Additive Model (GAM) using different information criteria, such as AIC, AICc, and BIC.
It helps prevent overfitting by choosing the appropriate degrees of freedom for smoother terms.
}
\keyword{internal}
