% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexOR-package.R
\docType{package}
\name{flexOR-package}
\alias{flexOR-package}
\title{flexOR: Flexible Odds Ratio Curves}
\description{
Provides flexible odds ratio curves that enable modeling non-linear relationships between continuous predictors and binary outcomes. This package facilitates a deeper understanding of the impact of each continuous predictor on the outcome by presenting results in terms of odds ratio (OR) curves based on splines. These curves allow for comparison against a specified reference value, aiding in the interpretation of the predictor's effect.
}
\references{
Hosmer, D. W. and Lemeshow, S. and Sturdivant, R. X. (2013).
 \emph{Applied Logistic Regression: Third Edition}, John Wiley and Sons Inc.,
 New York, NY.
 
 Royston, P. and Altman, D. G. and Sauerbrei, W. (2006).
 Dichotomizing continuous predictors in multiple regression: A bad idea.
 \emph{Statistics in Medicine}, \bold{25}(1), 127–141.
 \doi{10.1002/sim.2331}
 
 Hastie, T. J. and Tibshirani, R. J. (1990).
 \emph{Generalized Additive Models}, Chapman & Hall/CRC, New York, NY.
 
 Wood, S. N. (2017).
 \emph{Generalized Additive Models: An Introduction with R: Second Edition},
 Chapman & Hall/CRC, London, UK.
 
 Akaike, H. (1974). A new look at the statistical model identification.
 \emph{IEEE Transactions on Automatic Control}, \bold{19}(6), 716–723.
 \doi{10.1109/TAC.1974.1100705}
 
 Hurvich, C. M. and Simonoff, J. S. and Tsai, C. (1998).
 Smoothing parameter selection in nonparametric regression using an improved
 akaike information criterion.
 \emph{Journal of the Royal Statistical Society Series B: Statistical Methodology},
 \bold{60}(2), 271–293.
 \doi{10.1111/1467-9868.00125}
 
 Schwarz, G. (1978). Estimating the dimension of a model.
 \emph{Annals of Statistics}, \bold{6}(2), 461–464.
 \doi{10.1214/aos/1176344136}
 
 Cadarso-Suárez, C. and Meira-Machado, L. and Kneib, T. and Gude, F. (2010).
 Flexible hazard ratio curves for continuous predictors in multi-state models:
 an application to breast cancer data. \emph{Statistical Modelling}, \bold{10}(3), 291–314.
 \doi{10.1177/1471082X0801000303}
 
 Meira-Machado, L. and Cadarso-Suárez, C. and Gude, F. and Araújo, A. (2013).
 smoothHR: An R Package for Pointwise Nonparametric Estimation of Hazard Ratio
 Curves of Continuous Predictors.
 \emph{Computational and Mathematical Methods in Medicine}, \bold{2013}, 11 pages.
 \doi{10.1155/2013/745742}
 
 de Boor, C. (2001).
 \emph{A Practical Guide to Splines: Revised Edition}, Springer, New York, NY.
 
 Wood, S. N. and Pya, N. and Safken, B. (2016).
 Smoothing Parameter and Model Selection for General Smooth Models.
 \emph{Journal of the American Statistical Association}, \bold{111}(516), 1548-1563.
 \doi{10.1080/01621459.2016.1180986}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/martaaaa/flexOR}
  \item Report bugs at \url{https://github.com/martaaaa/flexOR/issues}
}

}
\author{
\strong{Maintainer}: Marta Azevedo \email{marta.vasconcelos4@gmail.com} (\href{https://orcid.org/0009-0003-4412-6991}{ORCID})

Authors:
\itemize{
  \item Luis Meira-Machado \email{lmachado@math.uminho.pt} (\href{https://orcid.org/0000-0002-8577-7665}{ORCID})
  \item Artur Araujo \email{artur.stat@gmail.com} (\href{https://orcid.org/0000-0003-1419-4208}{ORCID})
}

}
\keyword{package}
