% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floor_to.R
\name{floor_to}
\alias{floor_to}
\title{floor_to: Rounds a number down}
\usage{
floor_to(x, to = 1)
}
\arguments{
\item{x}{The number to be rounded.}

\item{to}{The multiple to which the number should be rounded down. Default is 1.}
}
\value{
The number rounded down to the specified multiple.
}
\description{
This function rounds a number down to the specified multiple.
}
\examples{
floor_to(7, 3) # Returns 6, as 6 is the largest multiple of 3 less than or equal to 7.
floor_to(5, 2) # Returns 4, as 4 is the largest multiple of 2 less than or equal to 5.

}
